requirements.txt
pyqt
qscintilla2
qdarkstyle
